+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Add new flag `buf generate --watch`. It tracks changes in proto files and automatically runs `buf generate`.
 - Add `clean` as a top-level option in `buf.gen.yaml`, matching the `buf generate --clean` flag. If
   set to true, this will delete the directories, jar files, or zip files set to `out` for each
   plugin.